recover.c
Scans a copy of a CompactFlash (CF) card (card.raw) for "deleted" JPEG files. Uses "isJPEG" function to identify the sygnature start of a JPEG file by analysing the first four bytes of 512 byte blocks. If a JPEG is found, it is copied to the current directory. No command-line arguments are required for this program as "card.raw" is hardcoded within.